// src/interfaces/api.interface.ts
export interface baseRes<D = {}> {
	success: boolean
	code: number
	message: string
	data: D //返回的数据字段
}

export interface PageData<D = {}> {
	tempGroup: any
	bindList: any
	filePreviewUrl: string
	fileSaveUrl: string
	currentPage: number
	pageNum: number
	pageSize: number
	totalPage: number
	count: number
	list: D[]
	fileHost: string
	helpVideo: {
		videoUrl: string
		coverUrl: string
	}
}
// ...
export interface userData {
	enabled: any
	description: string
	permissionName: string
	id: string
	username: string
	realName: string
	nickname: string
	avatar: string
	telephone: string
	email: string
	sex: number
	birthday: number
	signature: string
	accessToken: string
	userOrigin: number
	auth: boolean
	lastLoginIp: string
	lastLoginDate: number
	createTs: number
	type: string
	isShowAllMenus: boolean
	roleList?: number[]
}
export interface sceneListItem {
	sceneId: string
	sceneName: string
	thumbnail?: string
	srcId?: string
	srcType?: string
	sz?: string
	validSz?: string
	createTs: number
	createdBy: string
	createdName: string
	userId: string
	userName: string
	userPhone: string
	sceneStatus?: string
}

export interface getSortListItem {
	id: string
	createTs: string
	createdBy: string
	updateTs: string
	updatedBy: string
	categoryName: string
	categoryUrl: string
	categorySort: number
	deleted: string
}

export interface platformsItem {
	key: number | string
	value: string
}
// ...
export interface upFileItem {
	filePreviewUrl: string
	fileSaveUrl: string
	fileSize: number
}

export interface styleListItem {
	styleId: string
	name: string
	sort: number
}

export interface userListItem {
	birthday: string
	email: string
	id: string
	idCard: string
	nickname: string
	realName: string
	roleList: null
	sex: number
	state: number
	telephone: string
	userOrigin: null
	username: string
	name?: string
}
// ...
export interface authInfoData {
	authorities: Authority[]
	userRoles: UserRole[]
}

export interface Authority {
	id: number
	parentId: number
	action: string
	icon: string
	name: string
	url?: string
	sort: number
}

export interface UserRole {
	roleId: number
	roleName: string
}
// ...
export interface renderQueueItem {
	id: string
	tempId: string
	tempName: string
	thumb: string
	validSz: number
	userId: string
	username: string
	nickname: string
	renderStatus: number
	renderStatusDict: string
	sort: number
	topStatus: boolean
	sweeps: number
	createTs: string
}

export interface renderInfo {
	id: string
	nickname: string
	remark: string
	renderEndTime: number
	renderStartTime: number
	renderStatus: number
	sort: number
	tempId: string
	tempName: string
	thumb: string
	userId: string
	username: string
}