// src/interfaces/params.interface.ts
export interface PageParams {
	pageNum?: number
	pageSize?: number
	keyword?: string
}

export interface fineuploadParams {
	partIndex: number
	partSize: number
	totalParts: number
	totalFileSize: number
	filename: number
	uuid: number
	file: string
}
export interface sceneListParams {
	sceneId?: string
	sceneName?: string
	srcType?: string
}

export interface brokenLineParams {
	obj: string
	queryType: number
	startTimeStr: string
	endTimeStr: string
	topicId?: string
	tempType?: string
	timeType?: string
	timeStr?: string
	companyId?: string
}

export interface orderListParams {
	sceneId?: string
	sceneName?: string
	srcType?: string
}
export interface commentListParams {
	sceneId?: string
	sceneName?: string
	srcType?: string
}
export interface saveSceneParams {
	sceneName: string
	mark?: string
	panoramaInfo: upFileItem
	objInfo: upFileItem
}

export interface hallListParams {
	sceneId?: string
	sceneName?: string
	srcType?: string
}

export interface hallListParams {
	id?: string
	name?: string
	validSz?: number
	userId?: string
}

export interface tempListParams {
	userId?: string
	keyword: string
	companyId?: string
	topicId?: string
}

export interface editSceneParams {
	id?: string
	name?: string
	validSz?: number
	userId?: string
}